+++ tests/test_httpx_sync.py
@@ -6,24 +6,19 @@ from pytest_httpx import httpx_mock, HTTPXMock
 
 def test_httpx_mock_without_response(httpx_mock: HTTPXMock):
     with pytest.raises(Exception) as exception_info:
-        httpx.get("http://test_url")
+        with httpx.Client() as client:
+            client.get("http://test_url")
     assert (
         str(exception_info.value)
         == "No mock can be found for GET request on http://test_url."
     )
 
 
-@pytest.mark.xfail(
-    raises=AssertionError, reason="Unused responses should fail test case."
-)
-def test_httpx_mock_unused_response(httpx_mock: HTTPXMock):
-    httpx_mock.add_response("http://test_url")
-
-
 def test_httpx_mock_default_response(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url")
 
-    response = httpx.get("http://test_url")
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
     assert response.content == b""
     assert response.status_code == 200
     assert response.headers == httpx.Headers({})
@@ -33,25 +28,27 @@ def test_httpx_mock_default_response(httpx_mock: HTTPXMock):
 def test_httpx_mock_with_one_response(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", content=b"test content")
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content"
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+        assert response.content == b"test content"
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content"
+        response = client.get("http://test_url")
+        assert response.content == b"test content"
 
 
 def test_httpx_mock_with_many_responses(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url", content=b"test content 1")
     httpx_mock.add_response("http://test_url", content=b"test content 2")
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 1"
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+        assert response.content == b"test content 1"
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 2"
+        response = client.get("http://test_url")
+        assert response.content == b"test content 2"
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 2"
+        response = client.get("http://test_url")
+        assert response.content == b"test content 2"
 
 
 def test_httpx_mock_with_many_responses_methods(httpx_mock: HTTPXMock):
@@ -66,23 +63,24 @@ def test_httpx_mock_with_many_responses_methods(httpx_mock: HTTPXMock):
     )
     httpx_mock.add_response("http://test_url", method="HEAD", content=b"test content 6")
 
-    response = httpx.post("http://test_url")
-    assert response.content == b"test content 2"
+    with httpx.Client() as client:
+        response = client.post("http://test_url")
+        assert response.content == b"test content 2"
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 1"
+        response = client.get("http://test_url")
+        assert response.content == b"test content 1"
 
-    response = httpx.put("http://test_url")
-    assert response.content == b"test content 3"
+        response = client.put("http://test_url")
+        assert response.content == b"test content 3"
 
-    response = httpx.head("http://test_url")
-    assert response.content == b"test content 6"
+        response = client.head("http://test_url")
+        assert response.content == b"test content 6"
 
-    response = httpx.patch("http://test_url")
-    assert response.content == b"test content 5"
+        response = client.patch("http://test_url")
+        assert response.content == b"test content 5"
 
-    response = httpx.delete("http://test_url")
-    assert response.content == b"test content 4"
+        response = client.delete("http://test_url")
+        assert response.content == b"test content 4"
 
 
 def test_httpx_mock_with_many_responses_status_codes(httpx_mock: HTTPXMock):
@@ -105,29 +103,30 @@ def test_httpx_mock_with_many_responses_status_codes(httpx_mock: HTTPXMock):
         "http://test_url", method="HEAD", content=b"test content 6", status_code=500
     )
 
-    response = httpx.post("http://test_url")
-    assert response.content == b"test content 2"
-    assert response.status_code == 201
+    with httpx.Client() as client:
+        response = client.post("http://test_url")
+        assert response.content == b"test content 2"
+        assert response.status_code == 201
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 1"
-    assert response.status_code == 200
+        response = client.get("http://test_url")
+        assert response.content == b"test content 1"
+        assert response.status_code == 200
 
-    response = httpx.put("http://test_url")
-    assert response.content == b"test content 3"
-    assert response.status_code == 202
+        response = client.put("http://test_url")
+        assert response.content == b"test content 3"
+        assert response.status_code == 202
 
-    response = httpx.head("http://test_url")
-    assert response.content == b"test content 6"
-    assert response.status_code == 500
+        response = client.head("http://test_url")
+        assert response.content == b"test content 6"
+        assert response.status_code == 500
 
-    response = httpx.patch("http://test_url")
-    assert response.content == b"test content 5"
-    assert response.status_code == 404
+        response = client.patch("http://test_url")
+        assert response.content == b"test content 5"
+        assert response.status_code == 404
 
-    response = httpx.delete("http://test_url")
-    assert response.content == b"test content 4"
-    assert response.status_code == 303
+        response = client.delete("http://test_url")
+        assert response.content == b"test content 4"
+        assert response.status_code == 303
 
 
 def test_httpx_mock_with_many_responses_urls_str(httpx_mock: HTTPXMock):
@@ -150,23 +149,26 @@ def test_httpx_mock_with_many_responses_urls_str(httpx_mock: HTTPXMock):
         "http://test_url?param6=test", method="HEAD", content=b"test content 6"
     )
 
-    response = httpx.post(httpx.URL("http://test_url", params={"param2": "test"}))
-    assert response.content == b"test content 2"
+    with httpx.Client() as client:
+        response = client.post(httpx.URL("http://test_url", params={"param2": "test"}))
+        assert response.content == b"test content 2"
 
-    response = httpx.get(httpx.URL("http://test_url", params={"param1": "test"}))
-    assert response.content == b"test content 1"
+        response = client.get(httpx.URL("http://test_url", params={"param1": "test"}))
+        assert response.content == b"test content 1"
 
-    response = httpx.put(httpx.URL("http://test_url", params={"param3": "test"}))
-    assert response.content == b"test content 3"
+        response = client.put(httpx.URL("http://test_url", params={"param3": "test"}))
+        assert response.content == b"test content 3"
 
-    response = httpx.head(httpx.URL("http://test_url", params={"param6": "test"}))
-    assert response.content == b"test content 6"
+        response = client.head(httpx.URL("http://test_url", params={"param6": "test"}))
+        assert response.content == b"test content 6"
 
-    response = httpx.patch(httpx.URL("http://test_url", params={"param5": "test"}))
-    assert response.content == b"test content 5"
+        response = client.patch(httpx.URL("http://test_url", params={"param5": "test"}))
+        assert response.content == b"test content 5"
 
-    response = httpx.delete(httpx.URL("http://test_url", params={"param4": "test"}))
-    assert response.content == b"test content 4"
+        response = client.delete(
+            httpx.URL("http://test_url", params={"param4": "test"})
+        )
+        assert response.content == b"test content 4"
 
 
 def test_httpx_mock_with_many_responses_urls_instances(httpx_mock: HTTPXMock):
@@ -201,23 +203,24 @@ def test_httpx_mock_with_many_responses_urls_instances(httpx_mock: HTTPXMock):
         content=b"test content 6",
     )
 
-    response = httpx.post("http://test_url?param2=test")
-    assert response.content == b"test content 2"
+    with httpx.Client() as client:
+        response = client.post("http://test_url?param2=test")
+        assert response.content == b"test content 2"
 
-    response = httpx.get("http://test_url?param1=test")
-    assert response.content == b"test content 1"
+        response = client.get("http://test_url?param1=test")
+        assert response.content == b"test content 1"
 
-    response = httpx.put("http://test_url?param3=test")
-    assert response.content == b"test content 3"
+        response = client.put("http://test_url?param3=test")
+        assert response.content == b"test content 3"
 
-    response = httpx.head("http://test_url?param6=test")
-    assert response.content == b"test content 6"
+        response = client.head("http://test_url?param6=test")
+        assert response.content == b"test content 6"
 
-    response = httpx.patch("http://test_url?param5=test")
-    assert response.content == b"test content 5"
+        response = client.patch("http://test_url?param5=test")
+        assert response.content == b"test content 5"
 
-    response = httpx.delete("http://test_url?param4=test")
-    assert response.content == b"test content 4"
+        response = client.delete("http://test_url?param4=test")
+        assert response.content == b"test content 4"
 
 
 def test_httpx_mock_with_http_version_2(httpx_mock: HTTPXMock):
@@ -225,9 +228,10 @@ def test_httpx_mock_with_http_version_2(httpx_mock: HTTPXMock):
         "http://test_url", http_version="HTTP/2", content=b"test content 1"
     )
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 1"
-    assert response.http_version == "HTTP/2"
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+        assert response.content == b"test content 1"
+        assert response.http_version == "HTTP/2"
 
 
 def test_httpx_mock_with_headers(httpx_mock: HTTPXMock):
@@ -235,9 +239,10 @@ def test_httpx_mock_with_headers(httpx_mock: HTTPXMock):
         "http://test_url", content=b"test content 1", headers={"X-Test": "Test value"}
     )
 
-    response = httpx.get("http://test_url")
-    assert response.content == b"test content 1"
-    assert response.headers == httpx.Headers({"x-test": "Test value"})
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+        assert response.content == b"test content 1"
+        assert response.headers == httpx.Headers({"x-test": "Test value"})
 
 
 def test_httpx_mock_requests_retrieval(httpx_mock: HTTPXMock):
@@ -252,12 +257,13 @@ def test_httpx_mock_requests_retrieval(httpx_mock: HTTPXMock):
     )
     httpx_mock.add_response("http://test_url", method="HEAD", content=b"test content 6")
 
-    httpx.post("http://test_url", data=b"sent content 2")
-    httpx.get("http://test_url", headers={"X-TEST": "test header 1"})
-    httpx.put("http://test_url", data=b"sent content 3")
-    httpx.head("http://test_url")
-    httpx.patch("http://test_url", data=b"sent content 5")
-    httpx.delete("http://test_url", headers={"X-Test": "test header 4"})
+    with httpx.Client() as client:
+        client.post("http://test_url", data=b"sent content 2")
+        client.get("http://test_url", headers={"X-TEST": "test header 1"})
+        client.put("http://test_url", data=b"sent content 3")
+        client.head("http://test_url")
+        client.patch("http://test_url", data=b"sent content 5")
+        client.delete("http://test_url", headers={"X-Test": "test header 4"})
 
     assert (
         httpx_mock.get_request(httpx.URL("http://test_url"), "PATCH").read()
@@ -285,8 +291,9 @@ def test_httpx_mock_requests_retrieval(httpx_mock: HTTPXMock):
 def test_httpx_mock_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXMock):
     httpx_mock.add_response("http://test_url")
 
-    httpx.get("http://test_url", headers={"X-TEST": "test header 1"})
-    httpx.get("http://test_url", headers={"X-TEST": "test header 2"})
+    with httpx.Client() as client:
+        client.get("http://test_url", headers={"X-TEST": "test header 1"})
+        client.get("http://test_url", headers={"X-TEST": "test header 2"})
 
     assert (
         httpx_mock.get_request(httpx.URL("http://test_url")).headers["x-test"]
@@ -308,11 +315,12 @@ def test_httpx_mock_requests_json_body(httpx_mock: HTTPXMock):
     )
     httpx_mock.add_response("http://test_url", method="PUT", json="string value")
 
-    response = httpx.post("http://test_url")
-    assert response.json() == {"key 1": "value 1", "key 2": "value 2"}
+    with httpx.Client() as client:
+        response = client.post("http://test_url")
+        assert response.json() == {"key 1": "value 1", "key 2": "value 2"}
 
-    response = httpx.get("http://test_url")
-    assert response.json() == ["list content 1", "list content 2"]
+        response = client.get("http://test_url")
+        assert response.json() == ["list content 1", "list content 2"]
 
-    response = httpx.put("http://test_url")
-    assert response.json() == "string value"
+        response = client.put("http://test_url")
+        assert response.json() == "string value"